Move types from models to top-level types module
