+++ src/rpi_ai/functions.py
@@ -1,5 +1,5 @@
 from rpi_ai.function_calling.system_info import SystemInfo
-from rpi_ai.models.types import FunctionToolList
+from rpi_ai.types import FunctionToolList
 
 system_info = SystemInfo()
