+++ src/rpi_ai/main.py
@@ -10,7 +10,7 @@ from flask import Flask, Response, jsonify, request
 from rpi_ai.functions import FUNCTIONS
 from rpi_ai.models.chatbot import Chatbot
 from rpi_ai.models.logger import Logger
-from rpi_ai.models.types import AIConfigType
+from rpi_ai.types import AIConfigType
 
 logger = Logger(__name__)
