--- tests/rpi_ai/models/conftest.py
@@ -1,10 +1,0 @@
-from collections.abc import Generator
-from unittest.mock import MagicMock, patch
-
-import pytest
-
-
-@pytest.fixture
-def mock_extract_parts() -> Generator[MagicMock, None, None]:
-    with patch("rpi_ai.models.types.Message.extract_parts") as mock:
-        yield mock
